Revert "image dimensions type fix 3"
This reverts commit 2f193a4bea78e4e27ffbab00a0b624566e85f1ca.

+++ tasks/seq/vragov_i_gaussian_filter_vertical/src/filter.cpp
@@ -7,8 +7,8 @@
 bool vragov_i_gaussian_filter_vertical_seq::GaussianFilterTaskSequential::PreProcessingImpl() {
   // Init value for input and output
   unsigned int input_size = task_data->inputs_count[0];
-  x_ = static_cast<size_t>(task_data->inputs_count[1]);
-  y_ = static_cast<size_t>(task_data->inputs_count[2]);
+  x_ = static_cast<int>(task_data->inputs_count[1]);
+  y_ = static_cast<int>(task_data->inputs_count[2]);
   auto *in_ptr = reinterpret_cast<int *>(task_data->inputs[0]);
   input_ = std::vector<int>(in_ptr, in_ptr + input_size);
 
@@ -31,17 +31,17 @@ bool vragov_i_gaussian_filter_vertical_seq::GaussianFilterTaskSequential::RunImp
   if (input_.empty()) {
     return true;
   }
-  for (size_t i = 0; i < x_; i++) {
-    for (size_t j = 0; j < y_; j++) {
+  for (int i = 0; i < x_; i++) {
+    for (int j = 0; j < y_; j++) {
       double sum = 0.0;
       for (int k = -1; k <= 1; k++) {
-        int idx = static_cast<int>(j) + k;
+        int idx = j + k;
         if (idx < 0) {
           idx = 0;
-        } else if (idx >= static_cast<int>(y_)) {
-          idx = static_cast<int>(y_) - 1;
+        } else if (idx >= y_) {
+          idx = y_ - 1;
         }
-        sum += input_[(i * y_) + static_cast<size_t>(idx)] * kernel[k + 1];
+        sum += input_[(i * y_) + idx] * kernel[k + 1];
       }
       sum /= (sqrt(2.0 * pi) * 0.5);
       output_[(i * y_) + j] = static_cast<int>(std::round(sum));
